backend/requirements.txt
django
djangorestframework
pandas
django-cors-headers
channels
daphne
psycopg2-binary
python-dotenv
whitenoise
